+++ lib/src/twitter_api.dart
@@ -5,6 +5,7 @@
 // Project imports:
 import 'package:twitter_api_v2/src/client/client_context.dart';
 import 'package:twitter_api_v2/src/client/oauth_tokens.dart';
+import 'package:twitter_api_v2/src/service/spaces/spaces_service.dart';
 import 'package:twitter_api_v2/src/service/tweets/tweets_service.dart';
 import 'package:twitter_api_v2/src/service/twitter_service.dart';
 import 'package:twitter_api_v2/src/service/users/users_service.dart';
@@ -22,6 +23,9 @@ abstract class TwitterApi {
 
   /// Returns the users service.
   UsersService get usersService;
+
+  /// Returns the spaces service.
+  SpacesService get spacesService;
 }
 
 class _TwitterApi implements TwitterApi {
@@ -43,4 +47,7 @@ class _TwitterApi implements TwitterApi {
 
   @override
   UsersService get usersService => _twitterService.usersService;
+
+  @override
+  SpacesService get spacesService => _twitterService.spacesService;
 }