+++ lib/twitter_api_v2.dart
@@ -14,3 +14,5 @@ export 'package:twitter_api_v2/src/service/tweets/tweet_count_data.dart';
 export 'package:twitter_api_v2/src/service/tweets/tweet_count_meta.dart';
 export 'package:twitter_api_v2/src/service/users/user_data.dart';
 export 'package:twitter_api_v2/src/service/users/user_meta.dart';
+export 'package:twitter_api_v2/src/service/spaces/space_data.dart';
+export 'package:twitter_api_v2/src/service/spaces/space_meta.dart';